+++ packages/core/src/models/BaseModel.ts
@@ -1,4 +1,4 @@
-import { Store } from 'redux';
+import { AnyAction, Store } from 'redux';
 import { initModel, setCurrentModel, getModelName } from '../utils/model';
 import { NormalAction, IActionNormal } from '../actions/NormalAction';
 import { ComposeAction } from '../actions/ComposeAction';
@@ -9,6 +9,7 @@ import { ForgetRegisterError } from '../exceptions/ForgetRegisterError';
 import { NullReducerError } from '../exceptions/NullReducerError';
 import { storeHelper } from '../stores/StoreHelper';
 import { IActionPayload } from '../actions/BaseAction';
+import ACTION_TYPES from '../utils/actionType';
 
 export type FilterPersist<Data> = ((state: State<Data>) => StateReturn<Data>) | null;
 
@@ -61,6 +62,28 @@ export abstract class BaseModel<Data = null, RequestOption extends object = obje
    * ```
    */
   public static init = initModel;
+
+  /**
+   * Reset store and reassign initial state for each model.
+   * ```javascript
+   * logout().then(() => {
+   *   Model.resetStore();
+   * });
+   * ```
+   * You can keep current state by override method `keepStateOnReset()`
+   * ```javascript
+   * class TestModel extends Model {
+   *   protected keepOnResetStore() {
+   *     return true;
+   *   }
+   * }
+   * ```
+   */
+  public static resetStore(): AnyAction {
+    return storeHelper.dispatch({
+      type: ACTION_TYPES.reset,
+    });
+  }
 
   private readonly _name: string;
   private _action?: (() => IActionNormal<Data>) & NormalAction<Data, (state: State<Data>) => StateReturn<Data>, any, any>;
@@ -212,7 +235,7 @@ export abstract class BaseModel<Data = null, RequestOption extends object = obje
    * }
    *
    */
-  protected filterPersistData(): FilterPersist<Data> {
+  public/*protected*/ filterPersistData(): FilterPersist<Data> {
     return null;
   }
 
@@ -243,7 +266,7 @@ export abstract class BaseModel<Data = null, RequestOption extends object = obje
    * }
    * ```
    */
-  protected subscriptions(): Subscriptions<Data> {
+  public/*protected*/ subscriptions(): Subscriptions<Data> {
     return [];
   }
 
@@ -298,6 +321,14 @@ export abstract class BaseModel<Data = null, RequestOption extends object = obje
   ): void {}
 
   /**
+   * Keep the current state when resetStore() is invoked.
+   * @default false
+   */
+  public/*protected*/ keepOnResetStore(): boolean {
+    return false;
+  }
+
+  /**
    * The initial state for reducer.
    * When you enable persist to effect this model, the persist data will override it.
    */
@@ -308,8 +339,7 @@ export abstract class BaseModel<Data = null, RequestOption extends object = obje
       new BaseReducer(
         this.getReducerName(),
         this.initialState(),
-        this.subscriptions(),
-        this.filterPersistData()
+        this,
       ).createReducer()
     );
   }